+++ js/function.js
@@ -41,3 +41,23 @@ const getIntegerNumber = (string3) => {
 };
 
 getIntegerNumber(1.5);
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+function createIdGenerator () {
+  let lastGeneratedID = 0;
+
+  return () => {
+    lastGeneratedID += 1;
+    return lastGeneratedID;
+  };
+}
+
+export { getRandomInteger, getRandomArrayElement, createIdGenerator };
